Gulp file looking pretty good, basic partitioning done. Some TODOs to resolve before ready for commit.
