+++ example.exp
@@ -3,8 +3,8 @@ input-nostyle -NukeIFrameFilter-> -HTMLWriter-> case-2
 input-nostyle -StyleTokenizerFilter-> -StyleFilter-> -StyleDetokenizerFilter-> reduced* -NukeIFrameFilter-> -HTMLWriter-> case-3
 reduced* -StyleMinimizationFilter-> -NukeIFrameFilter-> -HTMLWriter-> case-4
 
-input: mobile-([^-]*).json
-input-nostyle: mobile-([^-]*)-nostyle.json
+input: test-([^-]*).json
+input-nostyle: test-([^-]*)-nostyle.json
 case-1: mobile-$1-[nuked].html
 case-2: mobile-$1-[nuked, nostyle].html
 case-3: mobile-$1-[nuked, compressed, nostyle].html
